Refactor constructor parameter data types of Shape types
Type conversion for smaller to bigger (in terms of memory) types is implicit (called widening) therefore int is acceptable to be used in constructors, even if the fields are declared in as floats.

+++ src/main/java/com/codecool/geometry/shapes/EquilateralTriangle.java
@@ -3,7 +3,7 @@ package com.codecool.geometry.shapes;
 public class EquilateralTriangle extends Shape {
     private final float a;
 
-    public EquilateralTriangle(float a, int id) {
+    public EquilateralTriangle(int a, int id) {
         super(id);
         this.a = a;
     }
